Roll back to alpha3.1
I plan to stay up to date with the latest libui release. Perhaps HEAD or master will be a bit ahead of that, but the latest tag will be for the latest release of libui.

+++ Sources/Combobox.swift
@@ -20,7 +20,7 @@ class Combobox: Control {
 			return Int(clibui.uiComboboxSelected(self.op))
 		}
 		set {
-			clibui.uiComboboxSetSelected(self.op, Int32(newValue))
+			clibui.uiComboboxSetSelected(self.op, newValue)
 		}
 	}
